mcp-auth/python/requirements.txt
mcpauth~=0.1.1
mcp~=1.12.3
PyJWT~=2.10.1
httpx~=0.27.0
pydantic~=2.11.7
python-dotenv~=1.0.0
uvicorn~=0.34.0
fastapi~=0.115.6
